+++ netnut.js
@@ -3,6 +3,7 @@
 var glob = require('glob');
 var argentum = require('argentum');
 var Netnut = require('./src/netnut.js');
+var lookup = require('./lib/lookup.js');
 var path = require('path');
 var yaml = require('yamljs');
 
@@ -23,6 +24,43 @@ var options = {
 };
 
 var netnut = new Netnut(options);
+
+var plugins = [];
+var root = lookup.sync(__dirname, 'package.json');
+if (root) {
+    var pack = require(root);
+
+    if (pack.config
+        && pack.config.netnut
+        && pack.config.netnut.plugins
+        && Array.isArray(pack.config.netnut.plugins)) {
+
+        pack.config.netnut.plugins.forEach(plugin => {
+            if (!~plugins.indexOf(plugin)) {
+                plugins.push(plugin);
+            }
+        });
+    }
+}
+
+if (Array.isArray(args.plugins)) {
+    args.plugins.forEach(plugin => {
+        if (!~plugins.indexOf(plugin)) {
+            plugins.push(plugin);
+        }
+    });
+}
+
+plugins.forEach(plugin => {
+    var mod;
+    if (plugin.includes('/')) {
+        mod = require(path.resolve(plugin));
+    } else {
+        mod = require(plugin);
+    }
+
+    mod(netnut, args);
+});
 
 netnut.run(
     args.host,