+++ changelog.md
@@ -1,3 +1,8 @@
+# v0.2.0
+
+* Add command line plugins loading.
+* Add `package.json` plugins loading.
+
 # v0.1.3
 
 * Fix ssh shell private key issue.